GtkScale: Add margin to the scale trough to recover the original style - SL #3983
This is the same thing that current Adwaita theme does in
gnome-themes-standard to make the trough smaller than the slider.

The margin value is hardcoded to get it right in the XO sugar-100
theme, but should be calculated later, when we fix the sugar-72 theme.
It looks like we will need different SVG images for that theme too.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>
Signed-off-by: Ajay Garg <ajay@activitycentral.com>